Update speedtest.sh
Added better logging of failures or discrepancies in the internet connection so the frontend can have an accurate view of the times there are actual failures instead of just throwing the data out and showing a better average result than is actually true. Moved last unnecessary log statement to entrypoint.sh as that is where the messaging should be done unless absolutely necessary.