Create settimedreload.js
Created settimedreload.js. This file is currently unused as I can't get my server to properly serve the file at this time. the refresh is applied due to the fact that this javascript lives inside the html now and it calculates the time necessary to refresh at 10 minutes past the hour to be in sync with the backend and avoid read/write collisions.