Update docker-compose.yml
Updated docker-compose.yml to remove the unnecessary env file from the frontend and changed the host port to something not being used on my system to do testing and data collection for debug purposes. Feel free to change this and the host directory to what you desire as your port and directory structure is likely very different from mine.